Started to decouple RedBasicAuth and RedBrowser.
Removed set_writeable() and the back and forth connection between
RedBrowser and RedBasicAuth. We now use a beforeMethod which will set
the right owner_ and take care of this.
This still needs more work, especially for the other methods and
RedBrowser is not the right place for this actually, too. It should go
into a RedDAVACL plugin.
Tried to clean up also some member variables of RedBasicAuth.
Moved profile loading from mod/cloud.php to RedBrowser, as it is only
needed there.